requirements.txt
langchain
langchain-core
langchain-community
langchain-experimental
langsmith
python-dotenv
langchain-openai
faiss-cpu